Split Video class into its own files
